+++ ChangeLog.md
@@ -49,6 +49,9 @@
 
 ### Language
 
+* **[breaking]** Makes the `Duplicate` `@value interface` from `lib/util` a
+  builtin and implements it for `Bool`, `Char`, `Float`, `Int`, and `String`.
+
 * **[fix]** Fixes checking of `defer`ed initialization in `scoped` blocks.
 
 * **[new]** Allows `present`, `require`, `strong`, and `reduce<#x,#y>` builtin